Migrate table query v2
