+++ front/components/agent_builder/capabilities/shared/SelectedDataSources.tsx
@@ -20,10 +20,8 @@ import { useKnowledgePageContext } from "@app/components/data_source_view/contex
 import type { DataSourceBuilderTreeItemType } from "@app/components/data_source_view/context/types";
 import { useMCPServerViewsContext } from "@app/components/shared/tools_picker/MCPServerViewsContext";
 import { useTheme } from "@app/components/sparkle/ThemeContext";
-import {
-  DATA_WAREHOUSE_SERVER_NAME,
-  TABLE_QUERY_V2_SERVER_NAME,
-} from "@app/lib/actions/mcp_internal_actions/constants";
+import { DATA_WAREHOUSE_SERVER_NAME } from "@app/lib/actions/mcp_internal_actions/constants";
+import { TABLE_QUERY_V2_SERVER_NAME } from "@app/lib/api/actions/servers/query_tables_v2/metadata";
 import { getConnectorProviderLogoWithFallback } from "@app/lib/connector_providers_ui";
 import { getVisualForContentNodeType } from "@app/lib/content_nodes";
 import {
