+++ front/lib/api/actions/servers/agent_router/metadata.ts
@@ -2,14 +2,13 @@ import type { JSONSchema7 as JSONSchema } from "json-schema";
 import { z } from "zod";
 import { zodToJsonSchema } from "zod-to-json-schema";
 
-import {
-  DEFAULT_AGENT_ROUTER_ACTION_DESCRIPTION,
-  DEFAULT_AGENT_ROUTER_ACTION_NAME,
-} from "@app/lib/actions/constants";
 import type { ServerMetadata } from "@app/lib/actions/mcp_internal_actions/tool_definition";
 import { createToolsRecord } from "@app/lib/actions/mcp_internal_actions/tool_definition";
 
-export const AGENT_ROUTER_TOOL_NAME = "agent_router" as const;
+export const AGENT_ROUTER_SERVER_NAME = "agent_router" as const;
+export const AGENT_ROUTER_ACTION_DESCRIPTION =
+  "Tools with access to the published agents of the workspace.";
+
 export const SUGGEST_AGENTS_TOOL_NAME = "suggest_agents_for_content" as const;
 
 export const AGENT_ROUTER_TOOLS_METADATA = createToolsRecord({
@@ -46,9 +45,9 @@ export const AGENT_ROUTER_TOOLS_METADATA = createToolsRecord({
 
 export const AGENT_ROUTER_SERVER = {
   serverInfo: {
-    name: DEFAULT_AGENT_ROUTER_ACTION_NAME,
+    name: AGENT_ROUTER_SERVER_NAME,
     version: "1.0.0",
-    description: DEFAULT_AGENT_ROUTER_ACTION_DESCRIPTION,
+    description: AGENT_ROUTER_ACTION_DESCRIPTION,
     authorization: null,
     icon: "ActionRobotIcon",
     documentationUrl: null,
