+++ front/lib/api/actions/servers/conversation_files/index.ts
@@ -3,7 +3,7 @@ import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import { makeInternalMCPServer } from "@app/lib/actions/mcp_internal_actions/utils";
 import { registerTool } from "@app/lib/actions/mcp_internal_actions/wrappers";
 import type { AgentLoopContextType } from "@app/lib/actions/types";
-import { CONVERSATION_FILES_TOOL_NAME } from "@app/lib/api/actions/servers/conversation_files/metadata";
+import { CONVERSATION_FILES_SERVER_NAME } from "@app/lib/api/actions/servers/conversation_files/metadata";
 import { TOOLS } from "@app/lib/api/actions/servers/conversation_files/tools";
 import type { Authenticator } from "@app/lib/auth";
 
@@ -15,7 +15,7 @@ function createServer(
 
   for (const tool of TOOLS) {
     registerTool(auth, agentLoopContext, server, tool, {
-      monitoringName: CONVERSATION_FILES_TOOL_NAME,
+      monitoringName: CONVERSATION_FILES_SERVER_NAME,
     });
   }
