+++ front/lib/api/actions/servers/web_search_browse/index.ts
@@ -1,10 +1,9 @@
 import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 
-import { DEFAULT_WEBSEARCH_ACTION_NAME } from "@app/lib/actions/constants";
 import { makeInternalMCPServer } from "@app/lib/actions/mcp_internal_actions/utils";
 import { registerTool } from "@app/lib/actions/mcp_internal_actions/wrappers";
 import type { AgentLoopContextType } from "@app/lib/actions/types";
-import { WEB_SEARCH_BROWSE_TOOL_NAME } from "@app/lib/api/actions/servers/web_search_browse/metadata";
+import { WEB_SEARCH_BROWSE_SERVER_NAME } from "@app/lib/api/actions/servers/web_search_browse/metadata";
 import { TOOLS } from "@app/lib/api/actions/servers/web_search_browse/tools";
 import type { Authenticator } from "@app/lib/auth";
 
@@ -12,11 +11,11 @@ function createServer(
   auth: Authenticator,
   agentLoopContext?: AgentLoopContextType
 ): McpServer {
-  const server = makeInternalMCPServer(DEFAULT_WEBSEARCH_ACTION_NAME);
+  const server = makeInternalMCPServer(WEB_SEARCH_BROWSE_SERVER_NAME);
 
   for (const tool of TOOLS) {
     registerTool(auth, agentLoopContext, server, tool, {
-      monitoringName: WEB_SEARCH_BROWSE_TOOL_NAME,
+      monitoringName: WEB_SEARCH_BROWSE_SERVER_NAME,
     });
   }
