+++ front/lib/api/actions/servers/web_search_browse/metadata.ts
@@ -2,14 +2,12 @@ import type { JSONSchema7 as JSONSchema } from "json-schema";
 import { z } from "zod";
 import { zodToJsonSchema } from "zod-to-json-schema";
 
-import {
-  DEFAULT_WEBSEARCH_ACTION_DESCRIPTION,
-  DEFAULT_WEBSEARCH_ACTION_NAME,
-} from "@app/lib/actions/constants";
 import type { ServerMetadata } from "@app/lib/actions/mcp_internal_actions/tool_definition";
 import { createToolsRecord } from "@app/lib/actions/mcp_internal_actions/tool_definition";
 
-export const WEB_SEARCH_BROWSE_TOOL_NAME = "web_search_browse" as const;
+export const WEB_SEARCH_BROWSE_SERVER_NAME = "web_search_&_browse" as const;
+export const WEB_SEARCH_BROWSE_ACTION_DESCRIPTION =
+  "Agent can search (Google) and retrieve information from specific websites.";
 
 export const WEB_SEARCH_BROWSE_TOOLS_METADATA = createToolsRecord({
   websearch: {
@@ -60,9 +58,9 @@ export const WEB_SEARCH_BROWSE_TOOLS_METADATA = createToolsRecord({
 
 export const WEB_SEARCH_BROWSE_SERVER = {
   serverInfo: {
-    name: DEFAULT_WEBSEARCH_ACTION_NAME,
+    name: WEB_SEARCH_BROWSE_SERVER_NAME,
     version: "1.0.0",
-    description: DEFAULT_WEBSEARCH_ACTION_DESCRIPTION,
+    description: WEB_SEARCH_BROWSE_ACTION_DESCRIPTION,
     authorization: null,
     icon: "ActionGlobeAltIcon" as const,
     documentationUrl: null,
